sve: top_bits
