+++ .github/workflows/main.yml
@@ -39,9 +39,9 @@ jobs:
         - { comp: gcc         , arch: aarch64 , opts: -Wno-psabi, status: 1 }
         - { comp: gcc         , arch: arm     , opts: -Wno-psabi, status: 1 }
         - { comp: gcc         , arch: ppc64   , opts: -Wno-psabi, status: 1 }
-        - { comp: gcc         , arch: sve128  , opts: -Wno-psabi, status: 0 }
-        - { comp: gcc         , arch: sve256  , opts: -Wno-psabi, status: 0 }
-        - { comp: gcc         , arch: sve512  , opts: -Wno-psabi, status: 0 }
+        - { comp: gcc         , arch: sve128  , opts: -Wno-psabi -DEVE_NO_PCH, status: 0 }
+        - { comp: gcc         , arch: sve256  , opts: -Wno-psabi -DEVE_NO_PCH, status: 0 }
+        - { comp: gcc         , arch: sve512  , opts: -Wno-psabi -DEVE_NO_PCH, status: 0 }
 
     steps:
       - name: Fetch current branch
