+++ doc/internals/dev_cmake.md
@@ -30,7 +30,7 @@ generator.
 | X86_avx2             | cmake .. -G Ninja -DCMAKE_TOOLCHAIN_FILE=../cmake/toolchain/clang.x86.cmake  -DEVE_OPTIONS='-mavx2'                     |
 | X86_fma              | cmake .. -G Ninja -DCMAKE_TOOLCHAIN_FILE=../cmake/toolchain/clang.x86.cmake  -DEVE_OPTIONS='-mavx2 -mfma'               |
 | X86_avx512           | cmake .. -G Ninja -DCMAKE_TOOLCHAIN_FILE=../cmake/toolchain/clang.x86.cmake  -DEVE_OPTIONS='-march=skylake-avx512'      |
-| X86_avx512_no_native | cmake .. -G Ninja -DCMAKE_TOOLCHAIN_FILE=../cmake/toolchain/clang.x86.sde.cmake  -DEVE_OPTIONS='-march=skylake-avx512'  |
+| X86_avx512_no_native | cmake .. -G Ninja -DCMAKE_TOOLCHAIN_FILE=../cmake/toolchain/clang.x86_sde.cmake  -DEVE_OPTIONS='-march=skylake-avx512'  |
 | Aarch64(arm-v8)      | cmake .. -G Ninja -DCMAKE_TOOLCHAIN_FILE=../cmake/toolchain/gcc.aarch64.cmake                                           |
 | Arm (arm-v7)         | cmake .. -G Ninja -DCMAKE_TOOLCHAIN_FILE=../cmake/toolchain/gcc.arm.cmake                                               |
 | wasm                 | emcmake cmake .. -G Ninja  -DCMAKE_TOOLCHAIN_FILE='../cmake/toolchain/emcc.cmake'                                       |
